loss almost aligned
